Split audio_codec into probe / decode / avio
The FFmpeg decode layer was ~2,626 lines in one file. `audio_codec.rs` becomes
`audio_codec/`: `mod.rs` keeps the `DecodedSink` trait, the shared error/audio
types, and the public entry points; `probe.rs` holds format detection, `decode.rs`
the streaming decode loop and sinks, and `avio.rs` the custom-I/O FFI plumbing
(`AvioContext`/`WriteAvioContext` and their unsafe callbacks). Inline tests move
to `tests.rs`. Verbatim relocation, no signature or behavior change.
