+++ bae-core/src/audio_codec/decode.rs
@@ -1,0 +1,965 @@
+//! Decoding any audio format to PCM: the in-memory `decode_audio` path
+//! (returns or streams interleaved i32 samples) and the `SparseBuffer`-backed
+//! `decode_audio_streaming` loop (pushes f32 samples to a `TrackSink`), plus
+//! the per-thread FFmpeg fatal-error counter that tracks decode failures.
+
+use super::avio::{
+    avio_read_callback, avio_seek_callback, streaming_avio_read_callback,
+    streaming_avio_seek_callback, AvioContext, StreamingAvioContext,
+};
+use super::{av_err_str, DecodedAudio, DecodedSink, StreamingDecodeError, AVIO_BUFFER_SIZE};
+use crate::playback::{SharedSparseBuffer, TrackSink};
+use std::cell::Cell;
+use std::os::raw::{c_int, c_void};
+use std::ptr;
+use std::sync::Arc;
+use tracing::{debug, info, warn};
+
+// Thread-local FFmpeg error counter for per-decode error tracking
+thread_local! {
+    // clippy::missing_const_for_thread_local fires spuriously on the Android target
+    // even though `= const { ... }` is already the const form; suppress the false positive.
+    #[allow(clippy::missing_const_for_thread_local)]
+    static FFMPEG_DECODE_ERRORS: Cell<u32> = const { Cell::new(0) };
+}
+
+/// Reset the thread-local FFmpeg error counter
+fn reset_ffmpeg_errors() {
+    FFMPEG_DECODE_ERRORS.with(|c| c.set(0));
+}
+
+/// Get current FFmpeg error count for this thread
+fn get_ffmpeg_errors() -> u32 {
+    FFMPEG_DECODE_ERRORS.with(|c| c.get())
+}
+
+/// The `va_list` parameter type of FFmpeg's log callback, as bindgen renders it
+/// per target ABI. Linux and Android share calling conventions: SysV x86_64 ->
+/// pointer-to-`__va_list_tag`; AAPCS64 aarch64 -> a 32-byte register-save struct
+/// (`__BindgenOpaqueArray<u64, 4>`). macOS/Windows keep the `c_char` pointer
+/// (Apple's arm64 ABI passes varargs on the stack, so its `va_list` is `char*`).
+/// The callback never reads `_vl`; this type only has to match the signature
+/// bindgen generated for `av_log_set_callback` on the target, or assignment
+/// below won't type-check.
+#[cfg(all(
+    any(target_os = "linux", target_os = "android"),
+    target_arch = "x86_64"
+))]
+type FfmpegVaList = *mut ffmpeg_sys_next::__va_list_tag;
+#[cfg(all(
+    any(target_os = "linux", target_os = "android"),
+    target_arch = "aarch64"
+))]
+type FfmpegVaList = ffmpeg_sys_next::__BindgenOpaqueArray<u64, 4>;
+#[cfg(not(all(
+    any(target_os = "linux", target_os = "android"),
+    any(target_arch = "x86_64", target_arch = "aarch64")
+)))]
+type FfmpegVaList = *mut std::ffi::c_char;
+
+/// Custom FFmpeg log callback that counts fatal errors per-thread.
+unsafe extern "C" fn ffmpeg_log_callback(
+    _avcl: *mut c_void,
+    level: c_int,
+    _fmt: *const std::ffi::c_char,
+    _vl: FfmpegVaList,
+) {
+    // Only count AV_LOG_FATAL (8) and AV_LOG_PANIC (0).
+    // AV_LOG_ERROR (16) includes recoverable sync errors during seeking.
+    if level <= 8 {
+        FFMPEG_DECODE_ERRORS.with(|c| c.set(c.get() + 1));
+    }
+}
+
+/// Install our custom FFmpeg log callback
+fn install_ffmpeg_log_callback() {
+    use std::sync::Once;
+    static INIT: Once = Once::new();
+    INIT.call_once(|| unsafe {
+        ffmpeg_sys_next::av_log_set_callback(Some(ffmpeg_log_callback));
+    });
+}
+
+/// Decode any audio format to PCM samples.
+///
+/// If start_ms/end_ms are provided, only that time range is decoded.
+/// Returns interleaved i32 samples.
+pub fn decode_audio(
+    data: &[u8],
+    start_sample: Option<u64>,
+    end_sample: Option<u64>,
+) -> Result<DecodedAudio, String> {
+    let mut sink = CollectingSink {
+        samples: Vec::new(),
+        format: None,
+    };
+    decode_audio_to_sink(data, start_sample, end_sample, &mut sink)?;
+    // `decode_audio_to_sink` returning Ok means the format was probed, so
+    // `on_format` ran before any samples.
+    let (sample_rate, channels, bits_per_sample) = sink
+        .format
+        .expect("decode produced samples without a format");
+    Ok(DecodedAudio {
+        samples: sink.samples,
+        sample_rate,
+        channels,
+        bits_per_sample,
+    })
+}
+
+/// The sink behind `decode_audio`: collects the whole decode into a `Vec` and
+/// records the probed format.
+struct CollectingSink {
+    samples: Vec<i32>,
+    format: Option<(u32, u32, u32)>,
+}
+
+impl DecodedSink for CollectingSink {
+    fn on_format(&mut self, sample_rate: u32, channels: u32, bits_per_sample: u32) {
+        self.format = Some((sample_rate, channels, bits_per_sample));
+    }
+    fn on_samples(&mut self, samples: &[i32]) {
+        self.samples.extend_from_slice(samples);
+    }
+}
+
+/// Stream-decode audio, pushing the format then each interleaved-i32 chunk into
+/// `sink` instead of returning a buffered `Vec`. Same window semantics as
+/// [`decode_audio`]; `start_sample`/`end_sample` trim the decoded range.
+pub fn decode_audio_to_sink(
+    data: &[u8],
+    start_sample: Option<u64>,
+    end_sample: Option<u64>,
+    sink: &mut dyn DecodedSink,
+) -> Result<(), String> {
+    // Safety: FFmpeg operations are contained within this function
+    unsafe { decode_audio_avio(data, start_sample, end_sample, sink) }
+}
+
+/// Internal AVIO-based decode implementation
+unsafe fn decode_audio_avio(
+    data: &[u8],
+    start_sample: Option<u64>,
+    end_sample: Option<u64>,
+    sink: &mut dyn DecodedSink,
+) -> Result<(), String> {
+    use ffmpeg_sys_next::*;
+
+    // Create our context for callbacks
+    let mut avio_ctx = Box::new(AvioContext {
+        data: data.as_ptr(),
+        size: data.len(),
+        pos: 0,
+    });
+
+    // Allocate AVIO buffer (FFmpeg will manage this)
+    let avio_buffer_size = AVIO_BUFFER_SIZE;
+    let avio_buffer = av_malloc(avio_buffer_size) as *mut u8;
+    if avio_buffer.is_null() {
+        return Err("Failed to allocate AVIO buffer".to_string());
+    }
+
+    // Create custom AVIO context
+    let avio = avio_alloc_context(
+        avio_buffer,
+        avio_buffer_size as c_int,
+        0, // read-only
+        avio_ctx.as_mut() as *mut AvioContext as *mut c_void,
+        Some(avio_read_callback),
+        None, // no write
+        Some(avio_seek_callback),
+    );
+    if avio.is_null() {
+        av_free(avio_buffer as *mut c_void);
+        return Err("Failed to create AVIO context".to_string());
+    }
+
+    // Create format context
+    let mut fmt_ctx = avformat_alloc_context();
+    if fmt_ctx.is_null() {
+        av_free(avio as *mut c_void);
+        return Err("Failed to allocate format context".to_string());
+    }
+    (*fmt_ctx).pb = avio;
+
+    // Open input (NULL filename since we're using custom I/O)
+    let ret = avformat_open_input(&mut fmt_ctx, ptr::null(), ptr::null_mut(), ptr::null_mut());
+    if ret < 0 {
+        avformat_free_context(fmt_ctx);
+        return Err(format!("Failed to open input: {}", av_err_str(ret)));
+    }
+
+    // Find stream info
+    let ret = avformat_find_stream_info(fmt_ctx, ptr::null_mut());
+    if ret < 0 {
+        avformat_close_input(&mut fmt_ctx);
+        return Err(format!("Failed to find stream info: {}", av_err_str(ret)));
+    }
+
+    // Find best audio stream
+    let stream_index = av_find_best_stream(
+        fmt_ctx,
+        AVMediaType::AVMEDIA_TYPE_AUDIO,
+        -1,
+        -1,
+        ptr::null_mut(),
+        0,
+    );
+    if stream_index < 0 {
+        avformat_close_input(&mut fmt_ctx);
+        return Err("No audio stream found".to_string());
+    }
+
+    let stream = *(*fmt_ctx).streams.add(stream_index as usize);
+    let codecpar = (*stream).codecpar;
+
+    // Find decoder
+    let codec = avcodec_find_decoder((*codecpar).codec_id);
+    if codec.is_null() {
+        avformat_close_input(&mut fmt_ctx);
+        return Err("Decoder not found".to_string());
+    }
+
+    // Allocate codec context
+    let codec_ctx = avcodec_alloc_context3(codec);
+    if codec_ctx.is_null() {
+        avformat_close_input(&mut fmt_ctx);
+        return Err("Failed to allocate codec context".to_string());
+    }
+
+    // Copy codec parameters
+    let ret = avcodec_parameters_to_context(codec_ctx, codecpar);
+    if ret < 0 {
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        return Err(format!("Failed to copy codec params: {}", av_err_str(ret)));
+    }
+
+    // Open codec
+    let ret = avcodec_open2(codec_ctx, codec, ptr::null_mut());
+    if ret < 0 {
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        return Err(format!("Failed to open codec: {}", av_err_str(ret)));
+    }
+
+    let sample_rate = (*codec_ctx).sample_rate as u32;
+    let channels = (*codecpar).ch_layout.nb_channels as u32;
+
+    // Determine bits per sample from format (for metadata only, actual extraction uses frame format)
+    let bits_per_sample = match (*codec_ctx).sample_fmt {
+        AVSampleFormat::AV_SAMPLE_FMT_U8 | AVSampleFormat::AV_SAMPLE_FMT_U8P => 8,
+        AVSampleFormat::AV_SAMPLE_FMT_S16 | AVSampleFormat::AV_SAMPLE_FMT_S16P => 16,
+        AVSampleFormat::AV_SAMPLE_FMT_S32 | AVSampleFormat::AV_SAMPLE_FMT_S32P => 32,
+        AVSampleFormat::AV_SAMPLE_FMT_FLT | AVSampleFormat::AV_SAMPLE_FMT_FLTP => 32,
+        AVSampleFormat::AV_SAMPLE_FMT_DBL | AVSampleFormat::AV_SAMPLE_FMT_DBLP => 64,
+        AVSampleFormat::AV_SAMPLE_FMT_S64 | AVSampleFormat::AV_SAMPLE_FMT_S64P => 64,
+        _ => 16,
+    };
+
+    sink.on_format(sample_rate, channels, bits_per_sample);
+
+    let time_base = (*stream).time_base;
+
+    // Seek to start position if specified (using stream time_base for exact positioning)
+    if let Some(sample_pos) = start_sample {
+        av_seek_frame(
+            fmt_ctx,
+            stream_index,
+            sample_pos as i64,
+            AVSEEK_FLAG_BACKWARD as c_int,
+        );
+    }
+
+    // Allocate frame and packet
+    let frame = av_frame_alloc();
+    let packet = av_packet_alloc();
+    if frame.is_null() || packet.is_null() {
+        if !frame.is_null() {
+            av_frame_free(&mut (frame as *mut _));
+        }
+        if !packet.is_null() {
+            av_packet_free(&mut (packet as *mut _));
+        }
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        return Err("Failed to allocate frame/packet".to_string());
+    }
+
+    let mut tracked_sample_pos: i64 = -1;
+
+    // Read and decode packets, trimming by sample position
+    let mut reached_end = false;
+    while av_read_frame(fmt_ctx, packet) >= 0 {
+        if (*packet).stream_index != stream_index {
+            av_packet_unref(packet);
+            continue;
+        }
+
+        let ret = avcodec_send_packet(codec_ctx, packet);
+        av_packet_unref(packet);
+
+        if ret < 0 {
+            continue;
+        }
+
+        while avcodec_receive_frame(codec_ctx, frame) >= 0 {
+            let num_samples = (*frame).nb_samples as usize;
+            let pts = (*frame).pts;
+
+            // Determine frame's sample position from PTS (with fallback tracking)
+            let frame_start = if pts != AV_NOPTS_VALUE {
+                let sample_pos = if time_base.num == 1 && time_base.den == sample_rate as c_int {
+                    pts
+                } else {
+                    (pts as f64 * time_base.num as f64 / time_base.den as f64 * sample_rate as f64)
+                        as i64
+                };
+                tracked_sample_pos = sample_pos;
+                sample_pos
+            } else if tracked_sample_pos >= 0 {
+                tracked_sample_pos
+            } else {
+                -1
+            };
+            let frame_end = frame_start + num_samples as i64;
+            if tracked_sample_pos >= 0 {
+                tracked_sample_pos = frame_end;
+            }
+
+            // Skip frames entirely before start
+            if let Some(start) = start_sample {
+                let start = start as i64;
+                if frame_start >= 0 && frame_end <= start {
+                    continue;
+                }
+            }
+
+            // Stop at end
+            if let Some(end) = end_sample {
+                let end = end as i64;
+                if frame_start >= 0 && frame_start >= end {
+                    reached_end = true;
+                    break;
+                }
+            }
+
+            let frame_samples_vec = extract_samples_from_raw_frame(frame, channels as usize);
+
+            // Trim start of frame
+            let skip_start = if let Some(start) = start_sample {
+                let start = start as i64;
+                if frame_start >= 0 && frame_start < start {
+                    let skip = (start - frame_start) as usize * channels as usize;
+                    skip.min(frame_samples_vec.len())
+                } else {
+                    0
+                }
+            } else {
+                0
+            };
+
+            // Trim end of frame
+            let take_end = if let Some(end) = end_sample {
+                let end = end as i64;
+                if frame_start >= 0 && frame_end > end {
+                    let keep = (end - frame_start) as usize * channels as usize;
+                    reached_end = true;
+                    keep.min(frame_samples_vec.len())
+                } else {
+                    frame_samples_vec.len()
+                }
+            } else {
+                frame_samples_vec.len()
+            };
+
+            if skip_start < take_end {
+                sink.on_samples(&frame_samples_vec[skip_start..take_end]);
+            }
+        }
+
+        if reached_end {
+            break;
+        }
+    }
+
+    // Flush decoder — only if we haven't reached end
+    if !reached_end {
+        avcodec_send_packet(codec_ctx, ptr::null());
+        while avcodec_receive_frame(codec_ctx, frame) >= 0 {
+            let frame_samples_vec = extract_samples_from_raw_frame(frame, channels as usize);
+            sink.on_samples(&frame_samples_vec);
+        }
+    }
+
+    // Cleanup
+    av_frame_free(&mut (frame as *mut _));
+    av_packet_free(&mut (packet as *mut _));
+    avcodec_free_context(&mut (codec_ctx as *mut _));
+    avformat_close_input(&mut fmt_ctx);
+    // Note: avformat_close_input frees the AVIO context and buffer
+
+    // Keep avio_ctx alive until here (prevent drop during FFmpeg operations)
+    drop(avio_ctx);
+
+    Ok(())
+}
+
+/// Extract samples from a raw AVFrame as i32
+unsafe fn extract_samples_from_raw_frame(
+    frame: *const ffmpeg_sys_next::AVFrame,
+    channels: usize,
+) -> Vec<i32> {
+    use ffmpeg_sys_next::{av_get_bytes_per_sample, AVSampleFormat};
+
+    let num_samples = (*frame).nb_samples as usize;
+    let mut samples = Vec::with_capacity(num_samples * channels);
+
+    // Get format info directly from the frame
+    let format: AVSampleFormat = std::mem::transmute((*frame).format);
+    let bytes_per_sample = av_get_bytes_per_sample(format);
+    if bytes_per_sample <= 0 {
+        warn!(
+            "av_get_bytes_per_sample returned {} for format {:?}, skipping frame",
+            bytes_per_sample,
+            (*frame).format
+        );
+        return Vec::new();
+    }
+    let actual_bytes_per_sample = bytes_per_sample as usize;
+
+    let is_float = matches!(
+        format,
+        AVSampleFormat::AV_SAMPLE_FMT_FLT
+            | AVSampleFormat::AV_SAMPLE_FMT_FLTP
+            | AVSampleFormat::AV_SAMPLE_FMT_DBL
+            | AVSampleFormat::AV_SAMPLE_FMT_DBLP
+    );
+
+    let is_planar = matches!(
+        format,
+        AVSampleFormat::AV_SAMPLE_FMT_U8P
+            | AVSampleFormat::AV_SAMPLE_FMT_S16P
+            | AVSampleFormat::AV_SAMPLE_FMT_S32P
+            | AVSampleFormat::AV_SAMPLE_FMT_FLTP
+            | AVSampleFormat::AV_SAMPLE_FMT_DBLP
+            | AVSampleFormat::AV_SAMPLE_FMT_S64P
+    );
+
+    if is_planar {
+        // Interleave from separate channel planes
+        for i in 0..num_samples {
+            for ch in 0..channels {
+                let plane = (*frame).data[ch] as *const u8;
+                if plane.is_null() {
+                    samples.push(0);
+                    continue;
+                }
+
+                let sample = read_sample(plane, i, actual_bytes_per_sample, is_float);
+                samples.push(sample);
+            }
+        }
+    } else {
+        // Packed format - all samples interleaved in plane 0
+        let data = (*frame).data[0] as *const u8;
+        if !data.is_null() {
+            for i in 0..(num_samples * channels) {
+                let sample = read_sample(data, i, actual_bytes_per_sample, is_float);
+                samples.push(sample);
+            }
+        }
+    }
+
+    samples
+}
+
+/// Read a single sample from raw bytes and convert to i32
+unsafe fn read_sample(
+    data: *const u8,
+    index: usize,
+    bytes_per_sample: usize,
+    is_float: bool,
+) -> i32 {
+    let offset = index * bytes_per_sample;
+    let ptr = data.add(offset);
+
+    if is_float {
+        let f = *(ptr as *const f32);
+        (f * i32::MAX as f32) as i32
+    } else {
+        match bytes_per_sample {
+            1 => (*(ptr as *const i8) as i32) * 256, // Scale 8-bit to 16-bit range
+            2 => *(ptr as *const i16) as i32,        // Keep 16-bit in native range
+            3 => {
+                // 24-bit little-endian, sign-extend to i32
+                let b = std::slice::from_raw_parts(ptr, 3);
+                let val = (b[0] as i32) | ((b[1] as i32) << 8) | ((b[2] as i32) << 16);
+                // Sign extend from 24-bit
+                if val & 0x800000 != 0 {
+                    val | 0xFF000000u32 as i32
+                } else {
+                    val
+                }
+            }
+            4 => *(ptr as *const i32),
+            _ => 0,
+        }
+    }
+}
+
+// =============================================================================
+// AVIO-based Streaming Decode
+// =============================================================================
+
+/// Decode audio from a SparseBuffer using FFmpeg's AVIO.
+///
+/// FFmpeg handles all frame boundary detection internally.
+/// Seektable is NOT needed - just feed bytes, get samples.
+/// Decode audio from a SparseBuffer, pushing f32 samples to a TrackSink.
+///
+/// The buffer holds the whole backing file. `seek_to_sample` jumps FFmpeg to the
+/// track's start; `start_at_sample` trims the lead-in FFmpeg outputs before the
+/// exact start (a frame may begin before it); `stop_at_sample` stops output at
+/// the track's end. `end_byte` is the track's end byte offset -- the read-ahead
+/// ceiling handed to the reader so the fill buffers the rest of this track;
+/// `None` keeps the whole file (per-track / last track).
+pub fn decode_audio_streaming(
+    buffer: SharedSparseBuffer,
+    sink: &mut TrackSink,
+    seek_to_sample: Option<u64>,
+    start_at_sample: Option<u64>,
+    stop_at_sample: Option<u64>,
+    end_byte: Option<u64>,
+    cancel_token: Arc<std::sync::atomic::AtomicBool>,
+) -> Result<(), StreamingDecodeError> {
+    install_ffmpeg_log_callback();
+    reset_ffmpeg_errors();
+
+    unsafe {
+        decode_audio_streaming_impl(
+            buffer,
+            sink,
+            seek_to_sample,
+            start_at_sample,
+            stop_at_sample,
+            end_byte,
+            cancel_token,
+        )
+    }
+}
+
+/// Internal AVIO-based streaming decode
+unsafe fn decode_audio_streaming_impl(
+    buffer: SharedSparseBuffer,
+    sink: &mut TrackSink,
+    seek_to_sample: Option<u64>,
+    start_at_sample: Option<u64>,
+    stop_at_sample: Option<u64>,
+    end_byte: Option<u64>,
+    cancel_token: Arc<std::sync::atomic::AtomicBool>,
+) -> Result<(), StreamingDecodeError> {
+    use ffmpeg_sys_next::*;
+
+    // Create streaming AVIO context. The reader's read-ahead ceiling is set
+    // after the seek below, once it sits at the track's start -- so the fill
+    // buffers the rest of the current track, not from byte 0 during probe.
+    let reader = buffer.new_reader_with_cancel(cancel_token.clone());
+    let cancel_status = cancel_token.clone();
+    let avio_ctx = Box::new(StreamingAvioContext {
+        reader: std::sync::Mutex::new(reader),
+        buffer: buffer.clone(),
+        cancel_token,
+    });
+    let avio_ctx_ptr = Box::into_raw(avio_ctx);
+
+    // Allocate AVIO buffer
+    let avio_buffer_size = AVIO_BUFFER_SIZE;
+    let avio_buffer = av_malloc(avio_buffer_size) as *mut u8;
+    if avio_buffer.is_null() {
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(
+            "Failed to allocate AVIO buffer",
+        ));
+    }
+
+    // Create custom AVIO context with seek support
+    let avio = avio_alloc_context(
+        avio_buffer,
+        avio_buffer_size as c_int,
+        0,
+        avio_ctx_ptr as *mut c_void,
+        Some(streaming_avio_read_callback),
+        None,
+        Some(streaming_avio_seek_callback),
+    );
+    if avio.is_null() {
+        av_free(avio_buffer as *mut c_void);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(
+            "Failed to create AVIO context",
+        ));
+    }
+
+    // Mark stream as seekable so avformat_seek_file works
+    (*avio).seekable = AVIO_SEEKABLE_NORMAL as c_int;
+
+    // Create format context
+    let mut fmt_ctx = avformat_alloc_context();
+    if fmt_ctx.is_null() {
+        av_free(avio as *mut c_void);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(
+            "Failed to allocate format context",
+        ));
+    }
+    (*fmt_ctx).pb = avio;
+
+    // Open input
+    let ret = avformat_open_input(&mut fmt_ctx, ptr::null(), ptr::null_mut(), ptr::null_mut());
+    if ret < 0 {
+        avformat_free_context(fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::input_error(
+            &cancel_status,
+            format!("Failed to open input: {}", av_err_str(ret)),
+        ));
+    }
+
+    // Find stream info
+    let ret = avformat_find_stream_info(fmt_ctx, ptr::null_mut());
+    if ret < 0 {
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::input_error(
+            &cancel_status,
+            format!("Failed to find stream info: {}", av_err_str(ret)),
+        ));
+    }
+
+    // Find audio stream
+    let stream_index = av_find_best_stream(
+        fmt_ctx,
+        AVMediaType::AVMEDIA_TYPE_AUDIO,
+        -1,
+        -1,
+        ptr::null_mut(),
+        0,
+    );
+    if stream_index < 0 {
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode("No audio stream found"));
+    }
+
+    let stream = *(*fmt_ctx).streams.add(stream_index as usize);
+    let codecpar = (*stream).codecpar;
+
+    // Find decoder
+    let codec = avcodec_find_decoder((*codecpar).codec_id);
+    if codec.is_null() {
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode("Decoder not found"));
+    }
+
+    // Allocate codec context
+    let codec_ctx = avcodec_alloc_context3(codec);
+    if codec_ctx.is_null() {
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(
+            "Failed to allocate codec context",
+        ));
+    }
+
+    // Copy codec parameters
+    let ret = avcodec_parameters_to_context(codec_ctx, codecpar);
+    if ret < 0 {
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(format!(
+            "Failed to copy codec params: {}",
+            av_err_str(ret)
+        )));
+    }
+
+    // Open codec
+    let ret = avcodec_open2(codec_ctx, codec, ptr::null_mut());
+    if ret < 0 {
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(format!(
+            "Failed to open codec: {}",
+            av_err_str(ret)
+        )));
+    }
+
+    let sample_rate = (*codec_ctx).sample_rate as u32;
+    let channels = (*codecpar).ch_layout.nb_channels as u32;
+
+    debug!("Streaming AVIO decoder: {}Hz, {}ch", sample_rate, channels);
+
+    // Set up SwrContext to convert any input format to packed f32
+    let in_ch_layout = (*codecpar).ch_layout;
+    let codec_ctx_format = (*codec_ctx).sample_fmt;
+    let mut swr_ctx: *mut SwrContext = std::ptr::null_mut();
+    let ret = swr_alloc_set_opts2(
+        &mut swr_ctx,
+        &in_ch_layout,
+        AVSampleFormat::AV_SAMPLE_FMT_FLT,
+        sample_rate as c_int,
+        &in_ch_layout,
+        codec_ctx_format,
+        sample_rate as c_int,
+        0,
+        std::ptr::null_mut(),
+    );
+    if ret < 0 || swr_ctx.is_null() {
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(format!(
+            "Failed to allocate SwrContext: {}",
+            av_err_str(ret)
+        )));
+    }
+    let ret = swr_init(swr_ctx);
+    if ret < 0 {
+        swr_free(&mut swr_ctx);
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(format!(
+            "Failed to init SwrContext: {}",
+            av_err_str(ret)
+        )));
+    }
+
+    // Seek to target sample position if requested
+    if let Some(sample_pos) = seek_to_sample {
+        let target_ts = sample_pos as i64;
+        let ret = avformat_seek_file(fmt_ctx, stream_index, i64::MIN, target_ts, target_ts, 0);
+        if ret < 0 {
+            warn!(
+                "avformat_seek_file failed ({}), decoding from start",
+                av_err_str(ret)
+            );
+        } else {
+            avcodec_flush_buffers(codec_ctx);
+        }
+    }
+
+    // The reader now sits at the track's start (after the seek above). Set its
+    // read-ahead ceiling -- the track's end byte, or the whole file when the
+    // track runs to EOF (a per-track file or an album's last track) -- so the
+    // fill buffers the rest of this track ahead of the playhead.
+    let ceiling = match end_byte {
+        Some(end) => end,
+        None => buffer.get_total_size(),
+    };
+    (*avio_ctx_ptr)
+        .reader
+        .lock()
+        .unwrap()
+        .set_readahead_ceiling(ceiling);
+
+    // Allocate frame and packet
+    let frame = av_frame_alloc();
+    let packet = av_packet_alloc();
+    if frame.is_null() || packet.is_null() {
+        if !frame.is_null() {
+            av_frame_free(&mut (frame as *mut _));
+        }
+        if !packet.is_null() {
+            av_packet_free(&mut (packet as *mut _));
+        }
+        swr_free(&mut swr_ctx);
+        avcodec_free_context(&mut (codec_ctx as *mut _));
+        avformat_close_input(&mut fmt_ctx);
+        let _ = Box::from_raw(avio_ctx_ptr);
+        return Err(StreamingDecodeError::decode(
+            "Failed to allocate frame/packet",
+        ));
+    }
+
+    let mut samples_output: u64 = 0;
+    let mut reached_stop = false;
+    let mut tracked_sample_pos: i64 = -1;
+
+    // Read and decode packets
+    while av_read_frame(fmt_ctx, packet) >= 0 {
+        // Check for cancellation
+        if sink.is_cancelled() || reached_stop {
+            av_packet_unref(packet);
+            break;
+        }
+
+        if (*packet).stream_index != stream_index {
+            av_packet_unref(packet);
+            continue;
+        }
+
+        let ret = avcodec_send_packet(codec_ctx, packet);
+        av_packet_unref(packet);
+
+        if ret < 0 {
+            continue;
+        }
+
+        while avcodec_receive_frame(codec_ctx, frame) >= 0 {
+            if sink.is_cancelled() {
+                break;
+            }
+
+            let num_samples = (*frame).nb_samples as usize;
+            let mut output_buf = vec![0.0f32; num_samples * channels as usize];
+            let out_ptr = output_buf.as_mut_ptr() as *mut u8;
+            swr_convert(
+                swr_ctx,
+                &out_ptr,
+                num_samples as c_int,
+                (*frame).extended_data as *const *const u8,
+                (*frame).nb_samples,
+            );
+
+            // Determine frame's sample position from PTS (with fallback tracking)
+            let time_base = (*stream).time_base;
+            let pts = (*frame).pts;
+            let frame_start_sample = if pts != ffmpeg_sys_next::AV_NOPTS_VALUE {
+                // PTS in stream time_base units — for audio this is typically sample number
+                let sample_pos = if time_base.num == 1 && time_base.den == sample_rate as c_int {
+                    pts
+                } else {
+                    // Convert PTS to sample position
+                    (pts as f64 * time_base.num as f64 / time_base.den as f64 * sample_rate as f64)
+                        as i64
+                };
+                tracked_sample_pos = sample_pos;
+                sample_pos
+            } else if tracked_sample_pos >= 0 {
+                tracked_sample_pos
+            } else {
+                -1
+            };
+            let frame_end_sample = frame_start_sample + num_samples as i64;
+            if tracked_sample_pos >= 0 {
+                tracked_sample_pos = frame_end_sample;
+            }
+
+            // Determine which portion of this frame to output
+            let mut skip_start: usize = 0;
+            let mut take_end: usize = output_buf.len();
+
+            // Trim start: skip samples before start_at_sample
+            if let Some(start) = start_at_sample {
+                let start = start as i64;
+                if frame_start_sample >= 0 && frame_end_sample <= start {
+                    continue;
+                }
+                if frame_start_sample >= 0 && frame_start_sample < start {
+                    let skip = (start - frame_start_sample) as usize * channels as usize;
+                    skip_start = skip.min(output_buf.len());
+                }
+            }
+
+            // Trim end: stop at stop_at_sample
+            if let Some(stop) = stop_at_sample {
+                let stop = stop as i64;
+                if frame_start_sample >= 0 && frame_start_sample >= stop {
+                    reached_stop = true;
+                    break;
+                }
+                if frame_start_sample >= 0 && frame_end_sample > stop {
+                    let keep = (stop - frame_start_sample) as usize * channels as usize;
+                    take_end = keep.min(output_buf.len());
+                    reached_stop = true;
+                }
+            }
+
+            let samples_to_output = if skip_start < take_end {
+                &output_buf[skip_start..take_end]
+            } else {
+                continue;
+            };
+
+            samples_output += samples_to_output.len() as u64;
+
+            if !samples_to_output.is_empty() && !push_samples_to_sink(sink, samples_to_output) {
+                break;
+            }
+        }
+    }
+
+    // Flush decoder — only if we haven't reached stop_at
+    if !reached_stop {
+        avcodec_send_packet(codec_ctx, ptr::null());
+        while avcodec_receive_frame(codec_ctx, frame) >= 0 {
+            if sink.is_cancelled() {
+                break;
+            }
+
+            let num_samples = (*frame).nb_samples as usize;
+            let mut output_buf = vec![0.0f32; num_samples * channels as usize];
+            let out_ptr = output_buf.as_mut_ptr() as *mut u8;
+            swr_convert(
+                swr_ctx,
+                &out_ptr,
+                num_samples as c_int,
+                (*frame).extended_data as *const *const u8,
+                (*frame).nb_samples,
+            );
+
+            samples_output += output_buf.len() as u64;
+
+            if !output_buf.is_empty() && !push_samples_to_sink(sink, &output_buf) {
+                break;
+            }
+        }
+    }
+
+    // Cleanup
+    av_frame_free(&mut (frame as *mut _));
+    av_packet_free(&mut (packet as *mut _));
+    swr_free(&mut swr_ctx);
+    avcodec_free_context(&mut (codec_ctx as *mut _));
+    avformat_close_input(&mut fmt_ctx);
+    let _ = Box::from_raw(avio_ctx_ptr);
+
+    // Record fatal error count (AV_LOG_FATAL and worse)
+    let error_count = get_ffmpeg_errors();
+    if error_count > 0 {
+        warn!(
+            "Streaming AVIO decode had {} fatal FFmpeg errors",
+            error_count
+        );
+    }
+    sink.set_decode_error_count(error_count);
+    sink.set_samples_decoded(samples_output);
+
+    if !sink.is_cancelled() {
+        sink.mark_finished();
+    }
+
+    info!(
+        "Streaming AVIO decode complete: {}Hz, {}ch, {} samples, {} fatal errors",
+        sample_rate, channels, samples_output, error_count
+    );
+
+    Ok(())
+}
+
+/// Push samples to sink in chunks, checking for cancellation between chunks.
+/// Returns `false` when the sink is cancelled, signaling the decode loop to exit.
+fn push_samples_to_sink(sink: &mut TrackSink, samples: &[f32]) -> bool {
+    const CHUNK_SIZE: usize = 8192;
+    for chunk in samples.chunks(CHUNK_SIZE) {
+        if sink.is_cancelled() {
+            return false;
+        }
+        sink.push_samples_blocking(chunk);
+    }
+    true
+}
